requirements.txt
requests==2.32.3
pandas==2.2.2
python-telegram-bot==21.5
gspread==6.0.2
google-auth==2.35.0
schedule==1.2.1
python-dotenv==1.0.1
